requirements/requirements.txt
# GraphQL
strawberry-graphql[debug-server,channels]==0.138.1
strawberry-graphql-django==0.7.1
strawberry-django-plus==1.28.6
gql==3.0.0
channels[daphne]==4.0.0
dj-database-url==0.5.0
django==4.1.3
django-pgschemas==0.11.0
psycopg2-binary==2.9.3
django-getenv==1.3.2
















